Added pundit task policy

+++ app/controllers/tasks_controller.rb
@@ -1,31 +1,37 @@
 # frozen_string_literal: true
 
 class TasksController < ApplicationController
+  after_action :verify_authorized, except: :index
+  after_action :verify_policy_scoped, only: :index
   before_action :authenticate_user_using_x_auth_token
   before_action :load_task, only: %i[show update destroy]
 
-  def show
-    @task_creator = User.find(@task.creator_id).name
-  end
-
   def index
-    tasks = Task.all
-    render status: :ok, json: { tasks: tasks }
+    @tasks = policy_scope(Task)
+    render status: :ok, json: { tasks: @tasks }
   end
 
   def create
     @task = Task.new(task_params.merge(creator_id: @current_user.id))
+    authorize @task
     if @task.save
-      render status: :ok, json: { notice: t("successfully_created", entity: "Task") }
+      render status: :ok,
+        json: { notice: t("successfully_created", entity: "Task") }
     else
       errors = @task.errors.full_messages.to_sentence
       render status: :unprocessable_entity, json: { error: errors }
     end
   end
 
+  def show
+    authorize @task
+    @task_creator = User.find(@task.creator_id).name
+  end
+
   def update
+    authorize @task
     if @task.update(task_params)
-      render status: :ok, json: { notice: "Successfully updated task." }
+      render status: :ok, json: {}
     else
       render status: :unprocessable_entity,
         json: { error: @task.errors.full_messages.to_sentence }
@@ -33,8 +39,9 @@ class TasksController < ApplicationController
   end
 
   def destroy
+    authorize @task
     if @task.destroy
-      render status: :ok, json: { notice: "Successfully deleted task." }
+      render status: :ok, json: {}
     else
       render status: :unprocessable_entity,
         json: { error: @task.errors.full_messages.to_sentence }
@@ -43,14 +50,14 @@ class TasksController < ApplicationController
 
   private
 
+    def task_params
+      params.require(:task).permit(:title, :user_id)
+    end
+
     def load_task
       @task = Task.find_by(slug: params[:slug])
       unless @task
         render status: :not_found, json: { error: t("task.not_found") }
       end
-    end
-
-    def task_params
-      params.require(:task).permit(:title, :user_id)
     end
 end
